move archives aggregate to module
